All functionalities working edition with two decimail points value.

+++ app/src/main/java/asc/msc/coursework/com/expensetracker/dto/Expense.java
@@ -1,8 +1,8 @@
 package asc.msc.coursework.com.expensetracker.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.ArrayList;
-import java.util.Date;
 
 /**
  * 
@@ -11,7 +11,7 @@ public class Expense extends Transaction implements Serializable {
 
     private int categoryId;
 
-    public Expense(String name, String comment, ArrayList<Integer> date, double value, int categoryId){
+    public Expense(String name, String comment, ArrayList<Integer> date, BigDecimal value, int categoryId){
         super(name,comment,date,value);
         this.categoryId =categoryId;
     }
